app/src/main/java/io/gogz/qperdiem/models/Rating.java
package io.gogz.qperdiem.models;


import androidx.room.Entity;
import androidx.room.PrimaryKey;

@Entity(tableName = "ratings")
public class Rating {

    @PrimaryKey private long ratingId;

    private String date;

    private Question question;

    public Rating(String date, Question question) {
        this.date = date;
        this.question = question;
    }

    public long getRatingId() {
        return ratingId;
    }

    public void setRatingId(long ratingId) {
        this.ratingId = ratingId;
    }

    public String getDate() {
        return date;
    }

    public void setDate(String date) {
        this.date = date;
    }

    public Question getQuestion() {
        return question;
    }

    public void setQuestion(Question question) {
        this.question = question;
    }
}
